refactor(rconsoleProvider): move lots. of the logic into smaller methds
