+++ src/rconsoleProvider.ts
@@ -38,155 +38,220 @@ export class MyComment implements vscode.Comment {
     }
 }
 
-function embelish_author(author: string) {
-    let embelished_author = author;
-    if (author === "user") {
-        embelished_author = "You";
-    }
-    if (author === "assistant") {
-        embelished_author = "🤖 Refact";
-    }
-    return embelished_author;
-}
+export class RefactConsoleProvider {
+    comment_controller: vscode.CommentController;
+    official_selection: vscode.Range;
+    attach_range: vscode.Range;
+    working_on_attach_code: string;
+    working_on_attach_filename: string;
+    code_snippet: string = "";
+    editor: vscode.TextEditor;
+    thread: vscode.CommentThread;
+    messages: rconsoleCommands.Messages;
 
-function message_to_comment(author: string, text: string) {
-    return new MyComment(text, vscode.CommentMode.Preview, new MyCommentAuthorInformation(embelish_author(author)));
-}
-
-// function format_messages(messages: rconsoleCommands.Messages) {
-//     return messages.filter(([role, _]) => {
-//         return role !== "context_file";
-//     }).map(([author, text]) => {
-//         return message_to_comment(author, text);
-//     });
-// }
-
-const disposable_commands: vscode.Disposable[] = [];
-const disposeCommands = () => {
-    disposable_commands.forEach(command => command.dispose());
-};
-const addCommand = (
-    cmd: string,
-    editor: vscode.TextEditor,
-    messages: rconsoleCommands.Messages,
-    update_thread_callback: rconsoleCommands.ThreadCallback,
-    end_thread_callback: rconsoleCommands.ThreadEndCallback
-) => {
-    const commandName = rconsoleCommands.createCommandName(cmd);
-    disposable_commands.push(
-        vscode.commands.registerCommand(commandName, () =>  {
-            activate_cmd(cmd, editor, messages, update_thread_callback, end_thread_callback);
-        })
-    );
-};
-
-const registerCommands = (
-    editor: vscode.TextEditor,
-    messages: rconsoleCommands.Messages,
-    update_thread_callback: rconsoleCommands.ThreadCallback,
-    end_thread_callback: rconsoleCommands.ThreadEndCallback
-) => {
-    disposeCommands();
-    Object.keys(rconsoleCommands.commands_available).forEach(cmd => {
-        addCommand(cmd, editor, messages, update_thread_callback, end_thread_callback);
-    });
-};
+    hint_debounce?: NodeJS.Timeout;
+    input_text = "";
+    hint_mode = true;
+    disposable_commands: vscode.Disposable[] = [];
 
 
-
-export function refact_console_close(): vscode.Uri|undefined
-{
-    if (global.comment_disposables) {
-        for (let d of global.comment_disposables) {
-            d.dispose();
+    static close_all_consoles(): vscode.Uri | undefined {
+        if (global.comment_disposables) {
+            for (let d of global.comment_disposables) {
+                d.dispose();
+            }
         }
+        global.comment_disposables = [];
+        let ret = global.comment_file_uri;
+        global.comment_file_uri = undefined;
+        return ret;
     }
-    disposeCommands();
-    global.comment_disposables = [];
-    let ret = global.comment_file_uri;
-    global.comment_file_uri = undefined;
-    return ret;
-}
 
-function update_context_file_with_insert_tag(messages: rconsoleCommands.Messages, editor: vscode.TextEditor): rconsoleCommands.Messages {
-    return messages.map(([type, text]) => {
-        if(type === "context_file") {
-            const [_selection, attach_range, working_on_attach_code, working_on_attach_filename, _code_snippet] = chatTab.attach_code_from_editor(editor, true);
-            return rconsoleCommands.initial_messages(working_on_attach_filename, working_on_attach_code, attach_range)[0];
-        }
-        return [type, text];
-    });
-}
-
-export async function open_refact_console_between_lines(editor: vscode.TextEditor)
-{
-    refact_console_close();
-    global.comment_file_uri = editor.document.uri;
-    let [official_selection, attach_range, working_on_attach_code, working_on_attach_filename, code_snippet] = chatTab.attach_code_from_editor(editor);
-    let cc = vscode.comments.createCommentController("refactai-test", "RefactAI Test Comments");
-    let next_line = Math.min(official_selection.end.line + 1, editor.document.lineCount - 1);
-    if (official_selection.isEmpty) {
-        next_line = official_selection.end.line;
+    static open_between_lines(editor: vscode.TextEditor) {
+        return new RefactConsoleProvider(editor);
     }
-    let comment_thread_range = new vscode.Range(next_line, 0, next_line, 0);
-    cc.commentingRangeProvider = {
-        provideCommentingRanges: (document: vscode.TextDocument) => {
-            return [comment_thread_range];
-        },
+
+    constructor(editor: vscode.TextEditor) {
+        RefactConsoleProvider.close_all_consoles();
+
+        this.editor = editor;
+        this.comment_controller =  vscode.comments.createCommentController("refactai-test", "RefactAI Test Comments");
+        global.comment_file_uri = editor.document.uri;
+
+        [
+            this.official_selection,
+            this.attach_range,
+            this.working_on_attach_code,
+            this.working_on_attach_filename,
+            this.code_snippet
+        ] = chatTab.attach_code_from_editor(editor);
+
+        this.dispose = this.dispose.bind(this);
+        this.handle_message_stream = this.handle_message_stream.bind(this);
+        this.handle_message_stream_end = this.handle_message_stream_end.bind(this);
+        this.handle_text_document_change = this.handle_text_document_change.bind(this);
+        this.handle_close_inline_chat = this.handle_close_inline_chat.bind(this);
+        this.handle_move_chat_to_sidebar = this.handle_move_chat_to_sidebar.bind(this);
+
+
+        this.thread = this.initialize_thread();
+        this.messages = this.initial_messages();
+
+        global.comment_disposables.push(this);
+        global.comment_disposables.push(
+            vscode.workspace.onDidChangeTextDocument(this.handle_text_document_change)
+        );
+        global.comment_disposables.push(
+            vscode.workspace.onDidCloseTextDocument(this.handle_text_document_close)
+        );
+
+        global.comment_disposables.push(
+            vscode.commands.registerCommand("refactaicmd.sendChatToSidebar", this.handle_move_chat_to_sidebar)
+        );
+
+        global.comment_disposables.push(
+            vscode.commands.registerCommand("refactaicmd.closeInlineChat", this.handle_close_inline_chat)
+        );
+
+        this.initial_message();
+
+    }
+
+    async initial_message() {
+        await vscode.commands.executeCommand('setContext', 'refactcx.runEsc', true);
+        // This trick puts cursor into the input box, possibly VS thinks the only use for
+        // the thread is to ask user if there are no messages. But then we add a message.
+        await new Promise(resolve => setTimeout(resolve, 100));
+        let [hint, author, _top1] = rconsoleCommands.get_hints(this.messages, "", this.official_selection);
+        const hint_comment = this.format_message(author, hint);
+        this.thread.comments = [hint_comment];
+    }
+
+    async handle_move_chat_to_sidebar() {
+        let question = "```\n" + this.code_snippet + "\n```\n\n" + this.input_text;
+        this.activate_chat(this.messages, question);
+    }
+
+    handle_close_inline_chat() {
+        global.comment_disposables.forEach(disposable => {
+            disposable.dispose();
+        });
+        global.comment_disposables = [];
+    }
+
+    remove_click_handlers_for_commands() {
+        this.disposable_commands.forEach(command => command.dispose());
+    }
+
+    add_click_handlers_for_commands() {
+        this.remove_click_handlers_for_commands();
+        Object.keys(rconsoleCommands.commands_available).forEach(cmd => {
+            const commandName = rconsoleCommands.createCommandName(cmd);
+            this.disposable_commands.push(
+                vscode.commands.registerCommand(commandName, () =>  {
+                    this.activate_cmd(cmd);
+                })
+            );
+        });
     };
-    cc.reactionHandler = (comment: vscode.Comment, reaction: vscode.CommentReaction): Thenable<void> => {
-        console.log("reactionHandler", comment, reaction);
-        return Promise.resolve();
-    };
 
-    let thread: vscode.CommentThread = cc.createCommentThread(
-        editor.document.uri,
-        comment_thread_range,
-        [],
-    );
-    let hint_mode = true;
-    thread.canReply = true;
-    thread.label = "Refact Console (F1)";
-    thread.collapsibleState = vscode.CommentThreadCollapsibleState.Expanded;
+    reset_thread() {
+        this.thread.comments = [
+            this.format_message("assistant", "Thinking...")
+        ];
+    }
 
-    global.comment_disposables.push(thread);
-    global.comment_disposables.push(cc);
-
-    let messages: [string, string][] = rconsoleCommands.initial_messages(working_on_attach_filename, working_on_attach_code, attach_range);
-
-    function messages_to_comments(clear = false) {
-        const assistant_messages = messages.filter(message => message[0] !== "context_file" && message[0] !== "user");
+    send_messages_to_thread() {
+        const assistant_messages = this.messages.filter(message => message[0] !== "context_file" && message[0] !== "user");
         const last_message = assistant_messages.slice(-1);
-        let new_comments = last_message.map(([author, message]) => message_to_comment(author, message));
+        let new_comments = last_message.map(([author, message]) => this.format_message(author, message));
 
-        if (new_comments.length === 0 || clear) {
-            new_comments = [
-                message_to_comment(embelish_author("assistant"), "Thinking...")
-            ];
+        if (new_comments.length === 0) {
+            this.reset_thread();
+        } else {
+            this.thread.comments = new_comments;
         }
-        thread.comments = new_comments;
     }
 
-    const update_thread_callback: rconsoleCommands.ThreadCallback = (author_role, answer) => {
-        const last = thread.comments.length > 0 ? thread.comments[thread.comments.length - 1] : null;
-        const comment = message_to_comment(embelish_author(author_role), answer);
+    format_message(author: string, text: string) {
+        let embellished_author = author;
+        if (author === "user") {
+            embellished_author = "You";
+        }
+        if (author === "assistant") {
+            embellished_author = "🤖 Refact";
+        }
+        const comment_author_info = new MyCommentAuthorInformation(embellished_author);
+        return new MyComment(text, vscode.CommentMode.Preview, comment_author_info);
+    }
+
+    initial_messages(): rconsoleCommands.Messages {
+        return rconsoleCommands.initial_messages(this.working_on_attach_filename, this.working_on_attach_code, this.attach_range);
+    }
+
+    dispose() {
+        console.log("console dispose");
+        console.log(this);
+        this.remove_click_handlers_for_commands();
+        this.thread.dispose();
+        this.comment_controller.dispose();
+    }
+
+    initialize_thread(): vscode.CommentThread {
+        const next_line = this.official_selection.isEmpty ? (
+            this.official_selection.end.line
+        ) : Math.min(
+            this.official_selection.end.line + 1,
+            this.editor.document.lineCount - 1
+        );
+
+        const comment_thread_range = new vscode.Range(next_line, 0, next_line, 0);
+        this.comment_controller.commentingRangeProvider = {
+            provideCommentingRanges: (document: vscode.TextDocument) => {
+                return [comment_thread_range];
+            },
+        };
+        this.comment_controller.reactionHandler = (comment: vscode.Comment, reaction: vscode.CommentReaction): Thenable<void> => {
+            console.log("reactionHandler", comment, reaction);
+            return Promise.resolve();
+        };
+
+        const thread = this.comment_controller.createCommentThread(
+            this.editor.document.uri,
+            comment_thread_range,
+            [],
+        );
+
+        thread.canReply = true;
+        thread.label = "Refact Console (F1)";
+        thread.collapsibleState = vscode.CommentThreadCollapsibleState.Expanded;
+
+        return thread;
+    }
+
+    handle_message_stream(author_role: string, answer: string) {
+        const last = this.thread.comments.length > 0 ? this.thread.comments[this.thread.comments.length - 1] : null;
+        const comment = this.format_message(author_role, answer);
+
         if(last instanceof MyComment && last.author.name === comment.author.name) {
-            const previouseComments = thread.comments.slice(0, -1);
-            thread.comments = [
-                ...previouseComments,
+            const previousComments = this.thread.comments.slice(0, -1);
+            this.thread.comments = [
+                ...previousComments,
                 comment
             ];
         } else {
-            thread.comments = [
-                ...thread.comments,
+            this.thread.comments = [
+                ...this.thread.comments,
                 comment,
             ];
         }
     };
 
-    const end_thread_callback = (response_messages: [string, string][]) => {
-        messages = response_messages;
-        messages_to_comments();
+    handle_message_stream_end(response_messages: rconsoleCommands.Messages) {
+        this.messages = response_messages;
+        console.log("end_thread_callback");
+        console.log({response_messages});
+        this.send_messages_to_thread();
         vscode.commands.executeCommand("setContext", "refactaicmd.openSidebarButtonEnabled", true);
         // "comments/commentThread/additionalActions": [
         // 	{
@@ -198,80 +263,84 @@ export async function open_refact_console_between_lines(editor: vscode.TextEdito
         // 		"group": "inline@2"
         // 	}
         // ]
-    };
-    let text = "";
-    let hint_debounce: NodeJS.Timeout|undefined;
+    }
 
-    let did1 = vscode.workspace.onDidChangeTextDocument(async e => {
-        console.log("onDidChangeTextDocument", e.document.uri, messages.length);
-        if (e.document.uri.scheme !== "comment") {
+    handle_text_document_close(e: vscode.TextDocument) {
+        if (e.uri.scheme !== "comment") {
             return;
         }
-        if (messages.length === 0) {
-            return;
-        }
-        text = e.document.getText();
-        if(text.startsWith("/")) {
-            registerCommands(editor, messages, update_thread_callback, end_thread_callback);
-        } else {
-            disposeCommands();
-        }
+        RefactConsoleProvider.close_all_consoles();
+    }
 
-        if (text.includes("\n")) {
-            let comment_editor = vscode.window.visibleTextEditors.find((e1) => {
-                return e1.document.uri === e.document.uri;
-            });
-            let first_line = text.split("\n")[0];
-            if (first_line.startsWith("/")) {
-                for (let cmd in rconsoleCommands.commands_available) {
-                    if (first_line.startsWith("/" + cmd)) {
-                        hint_mode = false;
-                        vscode.commands.executeCommand("setContext", "refactaicmd.openSidebarButtonEnabled", false);
-                        if (comment_editor) {
-                            await comment_editor.edit(edit => {
-                                edit.delete(new vscode.Range(0, 0, 1000, 0));
-                            });
-                        }
-                        // messages = [["Command", cmd]];
-                        // messages = [];
-                        messages_to_comments(true); // use this to remove chat
-                        const context_files = messages.filter(([type]) => type === "context_file");
-                        activate_cmd(cmd, editor, context_files, update_thread_callback, end_thread_callback);
-                        return;
+    get_messages_with_insert_tag_on_context_file() :rconsoleCommands.Messages {
+        return this.messages.map(([type, text]) => {
+            if(type === "context_file") {
+                const [_selection, attach_range, working_on_attach_code, working_on_attach_filename, _code_snippet] = chatTab.attach_code_from_editor(this.editor, true);
+                return rconsoleCommands.initial_messages(working_on_attach_filename, working_on_attach_code, attach_range)[0];
+            }
+            return [type, text];
+        });
+    }
+
+    async handle_user_pressed_enter(event: vscode.TextDocumentChangeEvent) {
+        // handle pressed enter
+        // active  chat also close the console
+        let comment_editor = vscode.window.visibleTextEditors.find((e1) => {
+            return e1.document.uri === event.document.uri;
+        });
+
+        let first_line = this.input_text.split("\n")[0];
+
+        if (first_line.startsWith("/")) {
+            for (let cmd in rconsoleCommands.commands_available) {
+                if (first_line.startsWith("/" + cmd)) { // maybe first_line.trim() === `/${cmd}`
+                    this.hint_mode = false;
+                    vscode.commands.executeCommand("setContext", "refactaicmd.openSidebarButtonEnabled", false);
+                    if (comment_editor) {
+                        await comment_editor.edit(edit => {
+                            edit.delete(new vscode.Range(0, 0, 1000, 0));
+                        });
                     }
+                    this.reset_thread();
+                    this.activate_cmd(cmd);
+                    return;
                 }
+            }
+        } else {
+            this.hint_mode = false;
+            if (this.code_snippet === "") {
+                const question = "Replace |INSERT-HERE| with the following:\n\n" + this.input_text;
+                const messages = this.get_messages_with_insert_tag_on_context_file();
+                this.activate_chat(messages, question);
+            } else if(this.messages.length > 1) {
+                this.activate_chat(this.messages, this.input_text);
             } else {
-                hint_mode = false;
-                let question;
-                const messages_with_insert_tag = update_context_file_with_insert_tag(messages, editor);
-                if (code_snippet === "") {
-                    question = "Replace |INSERT-HERE| with the following:\n\n" + text;
-                } else if(messages.length > 1) {
-                    question = text;
-                } else {
-                    question = "```\n" + code_snippet + "\n```\n\n" + text;
-                }
-                activate_chat(messages_with_insert_tag, question, editor);
+                const question = "```\n" + this.code_snippet + "\n```\n\n" + this.input_text;
+                this.activate_chat(this.messages, question);
             }
         }
+    }
 
-        let top1 = "";
-        if (hint_mode) {
-            let hint, author;
-            [hint, author, top1] = rconsoleCommands.get_hints(messages, text, official_selection);
-            let single_hint_comment = [message_to_comment(author, hint)];
-            if (hint_debounce) {
-                clearTimeout(hint_debounce);
+    async hints_and_magic_tabs(event: vscode.TextDocumentChangeEvent) {
+        const [hint, author, top1] = rconsoleCommands.get_hints(this.messages, this.input_text, this.official_selection);
+
+        if (this.hint_mode) {
+            if (this.hint_debounce) {
+                clearTimeout(this.hint_debounce);
             }
-            hint_debounce = setTimeout(async () => {
+            this.hint_debounce = setTimeout(async () => {
                 // this is a heavy operation, changes the layout and lags the UI
-                thread.comments = single_hint_comment;
+                this.thread.comments = [
+                    this.format_message(author, hint)
+                ]
             }, 200);
         }
+
         await vscode.commands.executeCommand('setContext', 'refactcx.runEsc', true);
-        if (top1 && text.match(/\/[a-zA-Z0-9_]+[\t ]+$/)) {
+
+        if (top1 && this.input_text.match(/\/[a-zA-Z0-9_]+[\t ]+$/)) {
             let comment_editor = vscode.window.visibleTextEditors.find((e1) => {
-                return e1.document.uri === e.document.uri;
+                return e1.document.uri === event.document.uri;
             });
             if (comment_editor) {
                 await comment_editor.edit(edit => {
@@ -281,79 +350,84 @@ export async function open_refact_console_between_lines(editor: vscode.TextEdito
             }
             return;
         }
-    });
-    let did2 = vscode.workspace.onDidCloseTextDocument(e => {
-        if (e.uri.scheme !== "comment") {
+    }
+
+    async handle_text_document_change(e: vscode.TextDocumentChangeEvent) {
+        console.log("onDidChangeTextDocument", e.document.uri, this.messages.length);
+
+        if (e.document.uri.scheme !== "comment" || this.messages.length === 0) {
             return;
         }
-        refact_console_close();
-    });
-    global.comment_disposables.push(did1);
-    global.comment_disposables.push(did2);
-    await vscode.commands.executeCommand('setContext', 'refactcx.runEsc', true);
-    function initial_message()
-    {
-        let [hint, author, _top1] = rconsoleCommands.get_hints(messages, "", official_selection);
-        const hint_comment = message_to_comment(author, hint);
-        thread.comments = [hint_comment];
+
+        this.input_text = e.document.getText();
+        this.hint_mode = this.input_text.startsWith("/");
+
+        if(this.hint_mode) {
+            this.add_click_handlers_for_commands();
+        } else {
+            this.remove_click_handlers_for_commands();
+        }
+
+
+        if (this.input_text.includes("\n")) {
+            return this.handle_user_pressed_enter(e);
+        }
+
+        await this.hints_and_magic_tabs(e);
     }
-    // This trick puts cursor into the input box, possibly VS thinks the only use for
-    // the thread is to ask user if there are no messages. But then we add a message.
-    await new Promise(resolve => setTimeout(resolve, 100));
-    initial_message();
 
-    global.comment_disposables.push(vscode.commands.registerCommand("refactaicmd.sendChatToSidebar", async (e) => {
-        let question = "```\n" + code_snippet + "\n```\n\n" + text;
-        activate_chat(messages, question, editor, false);
-    }));
 
-    global.comment_disposables.push(vscode.commands.registerCommand("refactaicmd.closeInlineChat", (...args) => {
-        global.comment_disposables.forEach(disposable => {
-            disposable.dispose();
-        });
-        global.comment_disposables = [];
-    }));
-}
+    activate_cmd(
+        cmd: string,
+    ) {
+        console.log(`activate_cmd refactaicmd.cmd_${cmd}`);
+        console.log({messages: this.messages});
+        const context_files = this.messages.filter(([type]) => type === "context_file");
+        vscode.commands.executeCommand("setContext", "refactaicmd.runningChat", true);
+        vscode.commands.executeCommand(
+            "refactaicmd.cmd_" + cmd,
+            this.editor.document.uri.toString(),
+            context_files,
+            this.handle_message_stream, // bind this
+            this.handle_message_stream_end // bind this
+        );
+    }
 
-function activate_cmd(cmd: string, editor: vscode.TextEditor, messages: rconsoleCommands.Messages, update_thread_callback: rconsoleCommands.ThreadCallback, end_thread_callback: rconsoleCommands.ThreadEndCallback)
-{
-    console.log(`activate_cmd refactaicmd.cmd_${cmd}`);
-    vscode.commands.executeCommand("setContext", "refactaicmd.runningChat", true);
-    vscode.commands.executeCommand("refactaicmd.cmd_" + cmd, editor.document.uri.toString(), messages, update_thread_callback, end_thread_callback);
-}
+    async activate_chat(
+        messages: [string, string][],
+        question: string,
+        new_question = true
+    ) {
+        console.log(`activate_chat question.length=${question.length}`);
+        RefactConsoleProvider.close_all_consoles();
+        await vscode.commands.executeCommand("refactai-toolbox.focus");
+        for (let i = 0; i < 10; i++) {
+            await new Promise(resolve => setTimeout(resolve, 100));
+            if (global.side_panel && global.side_panel._view) {
+                break;
+            }
+        }
+        let chat: chatTab.ChatTab|undefined = await sidebar.open_chat_tab(
+            question,
+            this.editor,
+            false,
+            "",
+            messages,
+            "");
+        if (!chat) {
+            return;
+        }
 
-async function activate_chat(messages: [string, string][], question: string, editor: vscode.TextEditor, new_question = true)
-{
-    console.log(`activate_chat question.length=${question.length}`);
-    refact_console_close();
-    await vscode.commands.executeCommand("refactai-toolbox.focus");
-    for (let i = 0; i < 10; i++) {
-        await new Promise(resolve => setTimeout(resolve, 100));
-        if (global.side_panel && global.side_panel._view) {
-            break;
+        if(new_question) {
+            await chat.post_question_and_communicate_answer(
+                question,
+                "",
+                "",
+                false,
+                messages,
+                );
+        } else {
+            await chat.chatHistoryProvider.save_messages_list(chat.chat_id, messages, "");
         }
     }
-    let chat: chatTab.ChatTab|undefined = await sidebar.open_chat_tab(
-        question,
-        editor,
-        false,
-        "",
-        messages,
-        "");
-    if (!chat) {
-        return;
-    }
-
-    if(new_question) {
-        await chat.post_question_and_communicate_answer(
-            question,
-            "",
-            "",
-            false,
-            messages,
-            );
-    } else {
-        await chat.chatHistoryProvider.save_messages_list(chat.chat_id, messages, "");
-    }
-
 }
